Moved all UI files into src/ui Added TODOs to toolbar.c Adjusted makefile accordingly Refactored include
