+++ Makefile
@@ -10,7 +10,7 @@ DEVFLAGS=-ggdb
 dev: bin/ligma-dev
 release: ligma-release
 
-bin/ligma-dev: src/main.c src/mainWindow.c src/toolbar.c src/viewport.c src/imagelist.c src/colorchooser.c
+bin/ligma-dev: src/main.c src/ui/mainWindow.c src/ui/toolbar.c src/ui/viewport.c src/ui/imagelist.c src/ui/colorchooser.c
 	${CC} $^ ${GTKFLAGS} ${DEVFLAGS} -o bin/ligma-dev ${GTKLIBS}
 
 clean:
